@sighphyre 2023-11-23
Hey @nathanmascitelli, thanks so much for this contribution, this is awesome!

Kinda curious though, I feel like the numbers here are a bit extreme, any chance you could share how many feature toggles you have in your Unleash instance? I'm just trying to find out if the SDK here is chewing through an excessive amount of memory or if you genuinely have a lot of toggles and this is just streamlining that