@nathanmascitelli 2023-11-23
@sighphyre we have 326 flags at the moment. I assumed this was a large number and that's why this hasn't been a problem for others. We are working on reducing the number of flags.