@sighphyre 2023-11-24
@nathanmascitelli Oh. Oh dear. No that's not a large number at all, that's well within reason for the computations (might be a little bit large for users to understand whats going on with their flags in their system though). For some context here, a similar sized feature set uses less than 15MB of RAM in our Edge server under load. And roughly a quarter of the numbers you're reporting for 40K toggles

Gosh time for some optimisation. Thank you so much for the information!