@nathanmascitelli 2023-11-24
@sighphyre sorry I'm not sure I follow your last statement. Are you saying its time for optimization of the dotnet client or my application? If my application, can you give some guidence on what optimizations you're thinking of? We are just calling `IsEnabled` on the `IUnleash` client created by `UnleashClientFactory`. Admittedly maybe we are calling it too much or something...