@sighphyre 2023-11-27
@nathanmascitelli Sorry! That was super unclear of me! I mean that the Unleash team probably needs to look this one over and see where the SDK is using these resources. Gut feel is that this isn't your application