@nathanmascitelli 2023-11-27
@sighphyre oh I would never count me out for doing something silly :). Jokes asside, thanks for clearing that up. I'm happy to help if I can.